Added methods get/set_memstore_merge_type in CompactionManager.
Memstore merge type can be either online or offline. If merge type is offline, then in order to
merge memstore with disk files we first flush memstore to disk to a new file, and then
merge this new file with other disk files. 
If merge is online, we don't flush memstore to disk, but merge it online with disk files.