Added a number of different variations of CompactionManager::merge_streams() function.
All these functions take as first argument a vector of streams to merge. Most new functions create by themselves a KVTDiskFile to which they store the output of merging input streams, and either return a pointer to this KVTFile, or append this pointer to a vector of KVTDiskFiles taken as argument.
There is also a variation that takes as a third argument a maximum file size and creates more than one disk files.