Added a read-write lock in DiskStore, so compactions/merges can be executed in parallel with get()s. The model is a "Single Writer - Multiple Readers", i.e. one thread that performs put()s and multiple threads that perform get()s.
A compaction manager must obtain a write-lock on DiskStore in order to modify it. Compaction manager only locks DiskStore at the very moment it needs to add a file to or delete a file from 'm_disk_files' vector. It doesn't need to get a read-lock to access DiskStore since it is the only class that can modify DiskStore.

A point_get() or range_get() must obtain a read-lock to DiskStore in order to read from disk files. Multiple threads can execute get()s in parallel.   

For performance reasons, a put() does not obtain a lock in MemStore before inserting a new <k,v>. This means that access to MemStore is not thread-safe: we cannot have put()s and get()s in parallel. So, in order to ensure consistency, RangeScanner point_get() and range_get() do not read values from memstore, only from diskstore. This of course means that, for example, if a key 'K' exists only in MemStore, get(K) will return false (i.e. "key does not exist").
Nevertheless, we focus on worst-case get() scenario where a get() needs to access disk, while also range_get()s need to access disk (every disk file, to be accurate) in any case.

Modified: 
 - DiskStore: added field 'm_rwlock'
 - GeometricCompactionManager, NomergeCompactionManager, RangemergeCompactionManager: get a write lock on DiskStore before modifying m_disk_files vector.
 - RangeScanner: get a read-lock on DiskStore while using disk files.