Fixed flush_mem bug in RangemergeCompactionManager.cpp
In main.cpp, we first set rangemerge flushmem size, and later
memstore max size. Because flushmem is set in set_flushmem() as:
   m_flushmem = min(flushmem, memstore->get_maxsize());
and because the actual memstore size is set later, the flush_mem
would *always* be set to the default memstore maxsize (100MB),
ignoring values given with -f, --rangemerge-flushmem args.