Cassandra algorithm works as in Apache Cassandra.
Previously, we merged the files of a level when we needed to create
a new file on that level and level already had K files ("lazy" approach).
Now, we merge the files of a level if, after the creation
of a new file on this level, the level has K files ("eager" approach).